edit preview style
ゴール位置の設定方法を変更
1.　最も近いfootstep位置を探す
2.　そのstepからN歩先の位置をゴール位置とする

+++ capturability/include/config.h
@@ -43,7 +43,7 @@ private:
 
   // parameter
   double timestep;
-  double preview;
+  int    preview;
 };
 
 } // namespace Capt
